refactor: migrate hand-rolled overlays and split ClientNonUIFeatures
