+++ src/app/features/settings/notifications/DeregisterPushNotifications.tsx
@@ -1,20 +1,7 @@
 import { useCallback, useState } from 'react';
-import FocusTrap from 'focus-trap-react';
-import {
-  Box,
-  Button,
-  color,
-  config,
-  Dialog,
-  Header,
-  IconButton,
-  Overlay,
-  OverlayBackdrop,
-  OverlayCenter,
-  Spinner,
-  Text,
-} from 'folds';
+import { Box, Button, color, config, Dialog, Header, IconButton, Spinner, Text } from 'folds';
 import { menuIcon, X } from '$components/icons/phosphor';
+import { ModalOverlay } from '$components/modal-overlay/ModalOverlay';
 import { useAtom } from 'jotai';
 import { useMatrixClient } from '../../../hooks/useMatrixClient';
 import { useClientConfig } from '../../../hooks/useClientConfig';
@@ -36,47 +23,38 @@ type ConfirmDeregisterDialogProps = {
 
 function ConfirmDeregisterDialog({ onClose, onConfirm, isLoading }: ConfirmDeregisterDialogProps) {
   return (
-    <Overlay open backdrop={<OverlayBackdrop />}>
-      <OverlayCenter>
-        <FocusTrap
-          focusTrapOptions={{
-            clickOutsideDeactivates: true,
-            onDeactivate: onClose,
-          }}
-        >
-          <Dialog variant="Surface">
-            <Header style={{ padding: `0 ${config.space.S400}` }} variant="Surface" size="500">
-              <Box grow="Yes">
-                <Text size="H4">Reset All Push Notifications</Text>
-              </Box>
-              <IconButton size="300" radii="300" onClick={onClose} disabled={isLoading}>
-                {menuIcon(X)}
-              </IconButton>
-            </Header>
-            <Box style={{ padding: config.space.S400 }} direction="Column" gap="400">
-              <Text>
-                This will remove push notifications from all your sessions and devices. This action
-                cannot be undone. Are you sure you want to continue?
-              </Text>
-              <Box direction="Column" gap="200" style={{ paddingTop: config.space.S200 }}>
-                <Button
-                  variant="Critical"
-                  fill="Solid"
-                  onClick={onConfirm}
-                  disabled={isLoading}
-                  before={isLoading && <Spinner size="100" variant="Critical" />}
-                >
-                  <Text size="B400">Reset All</Text>
-                </Button>
-                <Button variant="Secondary" fill="Soft" onClick={onClose} disabled={isLoading}>
-                  <Text size="B400">Cancel</Text>
-                </Button>
-              </Box>
-            </Box>
-          </Dialog>
-        </FocusTrap>
-      </OverlayCenter>
-    </Overlay>
+    <ModalOverlay requestClose={onClose}>
+      <Dialog variant="Surface">
+        <Header style={{ padding: `0 ${config.space.S400}` }} variant="Surface" size="500">
+          <Box grow="Yes">
+            <Text size="H4">Reset All Push Notifications</Text>
+          </Box>
+          <IconButton size="300" radii="300" onClick={onClose} disabled={isLoading}>
+            {menuIcon(X)}
+          </IconButton>
+        </Header>
+        <Box style={{ padding: config.space.S400 }} direction="Column" gap="400">
+          <Text>
+            This will remove push notifications from all your sessions and devices. This action
+            cannot be undone. Are you sure you want to continue?
+          </Text>
+          <Box direction="Column" gap="200" style={{ paddingTop: config.space.S200 }}>
+            <Button
+              variant="Critical"
+              fill="Solid"
+              onClick={onConfirm}
+              disabled={isLoading}
+              before={isLoading && <Spinner size="100" variant="Critical" />}
+            >
+              <Text size="B400">Reset All</Text>
+            </Button>
+            <Button variant="Secondary" fill="Soft" onClick={onClose} disabled={isLoading}>
+              <Text size="B400">Cancel</Text>
+            </Button>
+          </Box>
+        </Box>
+      </Dialog>
+    </ModalOverlay>
   );
 }
